Fix various compiler warnings and lint
This commit fixes various compiler warnings while building. There is
still one warning in the crates which compile in
`semantics::validate::SymContext` the `symtab` field is unused and
causing a warning. I didn't want to delete that because it might have
been future facing (although the struct isn't public). The
qasm3_bytecode crate fails to compile because a file is missing (I
assume a `git add` was missed on a recent push) so I didn't touch that
module yet.