Add basic CI configuration (#3)
This commit adds a basic initial CI configuration. This just adds a
single job that tests a build, runs `cargo fmt` and `cargo clippy` to
ensure consistent formatting and style, and then runs the unittests.
This provides coverage of the repo basics and we can expand and tweak
this over time as the library evolves.