Prepare crates for publishing
This commit updates the openqasm3_parser project for publishing crates.
Primarily this involves unifying the packaging metadata between all the
crates in the workspace. The other big change is all the crates and
library names are renamed to have a `oq3_` prefix with the exception of
qasm3_bytecode because it already had a descriptive name that was
unlikely to conflict with anything.