Reorganize sourcegen from crate to submodule in oq3_syntax
The sourcegen crate was published by rust-analyzer. But it was
broken, I pointed this out and they said publishing at all
was a mistake and removed it.

So we copied the crate into our workspace. But in fact it is
one file, used in one crate (oq3_syntax). So we move the
code to a module (a file) in oq3_syntax.