Define U gate before parsing
This adds a bogus gate definition for U, which is built in. This
appears as the first statement in the AST.

This is perhaps not the best way to handle the built in gate. But
it doesn't touch much code. It can be changed to something better without
too much trouble.