Make debugging in release build a compiler error
This commit adds macros ddbg and dprintln, which are defined in debug builds, but not
in release builds. So, if they appear in code they will result in a compile time
error when doing a release build. These are only added to the oq3_semantics crate.
They would have to be duplicated to use them in other crates. Or perhaps moved to the
lexer crate, which all others depend on, directly or indirectly.

Using these requires also #[cfg(debug_assertions)] when importing them.

I have seen an implementation of something similar that includes a definition of the
macros for release builds. This does not then require #[cfg(debug_assertions)] on
import. We could do somehting similar, by making ddbg expand to something that fails
to compile, say references an undefined variable.  But that would not be very robust.

There is a probably a standalone, registered, crate to do things like this. But I was unable to
find such a crate. Nor did I find online any reference to the basic problem this is meant to solve.