Make semdemo parse to AST use oq3_source_file
Before this commit, QASM_PATH is respected by `semdemo semantics` but not `semdemo parse`.

This commit allows files to be found in QASM_PATH when running `semdemo parse`.
This is needed when debugging and switching between `semdemo parse` and `semdemo semantics`.

For example, the following works with QASM_PATH
cargo run --example semdemo -- parse ex.qasm

Closes #41