Remove dependency on crate `source-span` (#49)
This dependency was listed in crates/oq3_source_file/Cargo.toml
But in fact it is not used. `source-span` depends on `termion`.
The latter does not work on Windows, so we don't want it.

It may be possible to disable the `termion` feature if for some reason
`source-span` turns out to be useful.