Parse argument to `measure` as gate operand
`GateOperand` is a qubit argument to gate call. And arguments to
`measure` have the same spec. Identifier, hardware identifier,
or indexed identifier. So we reuse the code from gate call.

The code in oq3_parser is a bit mysterious. I added some comments to
make it easier to locate next time, I hope.