Fix ast::Assignment
Replace `Name` with `(Name | IndexedIdentifier)` in the definition of
`AssignmentStmt` in openqasm3.ungram. This generates a method `indexed_identifier` implemented
for `AssignmentStmt` in nodes.rs. This allowed us to remove the handcoded method `indexed_identifier` in expr_ext.rs.
The latter exisited because I was going 100 mph implementing things and sticking the method in by hand was easier
than updating the ungram file.

The `SyntaxNode` underlying `AssignmentStmt` may have one or two `Expr`s. The last `Expr` is always the rhs.
So we take this into account when we retrieve it.

There is enough information in the ungrammar specification to write the method described above automatically.
But sourgen_ast.rs does not have the logic to this at present.