Update README after the first release
Now that the 0.0.1 release has been tagged and published on crates.io
this commit updates the README to reflect this a bit better. It adds
some basic tags to the README around the version on crates.io and the
MSRV and removes a section talking about a potential first release.