Bump version to 0.0.2
This tags as a release a commit that the Qiskit QASM 3 importer can
depend on. Starting with the commit that hte QASM 3 importer depends on,
two commits have been added on top: A change in the README, and this commit
bumping the version.

After this. The importer can depend on version 0.0.2