Try fixing patch version of clap: 4.4.15
In the previous commit, we set the version of clap to 4.4.
But in CI, an attempt is still made to install clap 4.5. No idea
why. But I was fairly confident that the attempt to fix CI failure
would fail somehow.

Let's see what fixing the version to 4.4.15 does in CI.

I should note that I am building this locally on linux using
rust 1.70, which is our desired MSRV. And the version of clap
being used is 4.4.15. Why does CI try to install 4.5, even
for ubuntu?