Fairly large refactoring of parsing assignment statements.
Previously, assignment statements were parsed in three separate places.  Now they are
parsed in one place. Some additional syntax error messages have been added.

Assignment statements are first parsed in the code that parses binary expressions at the
top level in a statement. Most binary expressions are then wrapped in `EXPR_STMT`.  But
assignments are filtered and are not wrapped. If an assignment is found in a non-top-level
postion a syntax error is logged.

`fn stmt` tries `expr_stmt` last, which is where we parse assignment statements.  We could
also try to parse an assignment earlier in `fn stmt`.  In this case, we would have to
abandon if it turns out not to be an assignment. This might cause performance degredation,
but I'm not at all sure.

If we keep going this route, we should probably organize `=`, `+=`, etc as variants in
an enum, similarly to how binary ops are handled. There is a fair amount of machinery for
binary ops. We might have to reproduce some of it.

Quite a bit of cruft was also removed. One r-a feature that we had carried around but not
used is now being used: `Restrictions` in expressions.rs.