Support declaring single hardware qubits
Statements like `qubit $1;` are supported.

This commit also undoes the previous commit (last commit before tagging 0.3.0) that
improved the error handling for attemted declaration of hardware qubits.

API CHANGE: new variant in asg::Stmt
The variant is  `Stmt::DeclareHardwareQubit(DeclareHardwareQubit)`

The struct `DeclareHardwareQubit` has method `name` to retrieve an instance of
`asg::HardwareQubit`. Note that `asg::HardwareQubit` is not new in this
commit.