Remove `struct Identifier` from asg.rs
As noted in a comment, having enum variants `Identifier(Identifier)`
where the field is a `struct Identifier` is not necessary, unless we
want to carry an extra copy of the name as a `String` for convenience.
This is a big price to pay for this convenience, which has not been used
for anything.

This commit removes `struct Identifier`. This also makes asg.rs more
consistent, since some of the variant already have the form `Identifier(SymbolIdResult)`