Remove unused run_rustfmt.sh (#207)
I added this before publishing in order to run rustfmt on only select portions of the code
while I vetted the results. For example sometimes comments are moved to places that change
their intended purpose. Then Matthew formatted the entire code base and made a PR. So I
thought, ok we can take the plunge.  But I neglected to remove this script.