Add tests of the example code in the OQ3 spec
This is the start of a big project. Especially becaue a good amount of the
code fails. This is in part because we have been increasing the amount of
semantic analysis. This is good because less incorrect code is silently
allowed. However the are many false positives.

This exercise is a good way to identify and organize features that
are not supported or that have bugs.

This commit covers all of "comments.rst" and 16 percent of the file
"types.rst".