Improvements to `SymbolTable::gates` function (#217)
* Filter U gate from output of SymbolTable::gates

The Qiskit importer otherwise needs to filter this gate out.
Other consumers probably would do so as well. If it is useful in
antother context, we can add an option to include the U gate

Review comment: https://github.com/Qiskit/qiskit/pull/12087#discussion_r1586950430

* Use impl trait for SymbolTable::gate return type
So we return an iterator, not a `Vec`.