Support prefixes 0X, 0O, 0B for hex, octal, binary literals
Closes #226

Previously, literals with prefix `0x` were parsed and analyzed correctly. But literals
with prefix `0X` were analyzed as value `0`. This is now fixed. Prefixes `0B` and `0O`
were broken in the same way and are now fixed.

This fix also uncovers a bug that will be opened and fixed separately: Illegal (nonsense)
integer literals, with prefixes such as 0Q will be analyzed as valid, decimal,
literals. Instead they should cause an error to be logged. A lex, or possibly parse,
error.