Update dependency tree and raise MSRV to 1.79
This commit updates the dependency tree to pull in the latest versions
of all the dependencies which were now out of date. The exception is
dev dependencies aren't update and also ra_ap_limit is held to 0.0.229
because newer releases require rust >= 1.80.0 which is higher than
Qiskit >2.0's MSRV of 1.79. Accordingly this commit updates the MSRV to
1.79 to make sure there aren't any compatibility issues with older
versions of Rust.