Add tests for code examples in OQ3.1 spec and fix bug found (#243)
The idea is to check that all of the code examples in the spec are parsed and analyzed correctly. But first, we aim for the easier goal of checking that no errors are reported. Three examples in subroutines.rst have been added.

One of these revealed a bug. A for statement with a single statement for the body, rather than a block was not supported, causing a panic when attempting to parse.

This PR fixes this bug.

Co-authored-by: John Lapeyre <1969884+jlapeyre@users.noreply.github.com>