Log a "semantic error" when an `include` statement fails. (#247)
* Log a "semantic error" when an `include` statement fails.

Previously: When parsing and analyzing OQ3 source, one might encounter an
`include` statement. In this case we read the included file and parse and analyze it.
If the file cannot be read for some reason, for example it does not exist, the parser panics.

With this commit/PR: A few data fields have been modified by wrapped in `Option`. And a
field has been added to the struct that manage parsing source files. This field is a
a struct giving information on the io error, and is wrapped in Option. Then in a call
to `syntax_to_semantics::syntax_to_semantic` these Options are checked. Trying to parse
the syntax tree for the file that could not be read is avoided. And a "semantic error" is
logged. The syntax element for the include is included in the error struct stored with the
source file info. This is used by the error reporting system to report the character range
in the source file at which the failed include occurred.

For example, parsing/analyzing this code
```
include "stdgates2.inc";

qubit q;
h q;
```
results in
```
[03] Error: FileNotFound
   ╭─[ /home/joe/qiskit/openqasm3_parser/crates/oq3_semantics/examples/qasm/noinclude.qasm:1:9 ]
 1 │include "stdgates2.inc";
   │               ╰──────── Near this point

[03] Error: UndefGateError
   ╭─[ /home/joe/qiskit/openqasm3_parser/crates/oq3_semantics/examples/qasm/noinclude.qasm:4:1 ]
 4 │h q;
   │╰─ Near this point
```
Note that parsing continued past the failed file read. And a subsequent error, also the result of the failed read, is reported as well.

Closes #74

* Simplify borrowing with getter

* Use borrow and explicit clone

* Add comment on borrow

* Simplify logic (or shorten logic)

* format

---------

Co-authored-by: John Lapeyre <1969884+jlapeyre@users.noreply.github.com>