Replace nth(0) by current() and remove dead code
I commented out the dead code. The only suggested use is for contextual keywords, which OQ3
does not have.

I replace one used of nth(0) with current(). A similar optimization is done in oq3_lexer
(and was done in the rust lexer. I don't know why these were not made before in the parser.