Add "invalid" qasm source tests from the reference parser
These uncover some bugs in the parser. One file causes parsing to hang.
This qasm file is present, but disabled by adding a suffix to it.
The parser panics on one or two of the files.

The file tokens.qasm is supposed produce at least one invalid (lexer) tokens
per line. In fact most of these are succesfully lexed.