Add several snippets from reference parser to tests (#269)
* Add snippet cal2.qasm that produces an invalid lexeme

I copied a calibration body from the snippet defcal.qasm to cal statement. This body has a single double quote. The lexer fails to produce a valid token when trying to parse this as a string. The lexer records this error in a field of a variant of Token.

* Add header.qasm taken from reference parser

lexing, parsing, analysis all succeed.

I updated the name of the standard gates library.

* Add "invalid" qasm source tests from the reference parser

These uncover some bugs in the parser. One file causes parsing to hang.
This qasm file is present, but disabled by adding a suffix to it.
The parser panics on one or two of the files.

The file tokens.qasm is supposed to produce at least one invalid (lexer) tokens
per line. In fact most of these are successfully lexed.

* Add note about parser hanging on a snippet

This causes the parser to hang.
snippets/invalid/statements/declarations.qasm.parser.hangs

The extra suffix prevents it from being lexed/parsed. It is tagged with "skip".
But the runner.rs is broken in this regard and tries to parse it anyway.

* Fix, improve runner.rs

Snippet test runner was not skipping stages tagged "skip".
This is needed for tests that hang the parser.