Refactor detecting end token in parsing lists of items
This replaces some clumsy code, with more efficient and cleaner code.
Various types, "flavors" of lists are parsed with a single function.
Each kind of list has a token or token that signifys no more items when
encountered. This commit factors this into a single function and removes
some redundancy.