Refactor detecting end token in parsing lists of items (#280)
This replaces some clumsy code with more efficient and cleaner code.

Various types-- "flavors" --of lists are parsed with a single function. The flavor is sent as a parameter to this function so that particulars may be treated correctly. Furthermore, each kind of list has a token or token that signifies no more items when it is encountered. This commit factors this check for end tokens into a single function and removes some redundancy.