Don't run stage tagged "panic" on snippet
Previously, the state was run, and the panic trapped.
We are not making use of the information resulting from the
panic. The simplest reasonable thing to do is not run stages
tagged "panic".

We leave this tag in place in order to say *why* we are skipping
the stage. The idea is to fix these so that we can do a proper skip,
and eventually, tag with ok.