Don't run stage tagged "panic" on snippet (#281)
Previously, the stage was run, and the panic trapped. We are not making use of the information resulting from the panic. The simplest
reasonable thing to do is not run stages tagged "panic".

We keep this tag as an option in order to say why we are skipping the 
stage. The idea is to fix these so that we can do a proper skip, and
eventually, tag with ok.